KUL_dcm2bids
Check if task-*_events.tsv exists, and only then copy it.

+++ KUL_dcm2bids.sh
@@ -560,8 +560,11 @@ rm -rf $bids_output/tmp_dcm2bids
 
 # copying task based events.tsv to BIDS directory
 if [ $events_flag -eq 1 ]; then
-    kul_e2cl "Copying task based events.tsv to BIDS directory" $log
-    cp Study_config/task-*_events.tsv $bids_output
+    test_events_exist=$(ls -l Study_config/task-*_events.tsv | grep "No such file")
+    if [ $test_events_exist = "" ]; then
+        kul_e2cl "Copying task based events.tsv to BIDS directory" $log
+        cp Study_config/task-*_events.tsv $bids_output
+    fi  
 fi
 
 # clean up
